requirements.txt
#フレームワーク
fastapi==0.128.0

Jinja2==3.1.6

pydantic==2.12.5

#データベース
SQLAlchemy==2.0.46

#ASGIサーバー
uvicorn==0.40.0

#gemini AI
google-genai>=0.3,<1.0

#.envファイル読み込み
python-dotenv==1.2.1
